# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gaia API",
    "description": "Gaia Large Knowledge Model API - manages knowledge graph operations",
    "version": "3.0.0",
    "contact": {
      "name": "Gaia API Support",
      "url": "https://github.com/SiliconEinstein/Gaia/issues"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {"url": "http://localhost:8000", "description": "Local development server"}
  ],
  "paths": {
...
    "/nodes": {
      "get": {
        "summary": "List nodes (paginated)",
        "description": "Get paginated list of nodes with optional type filtering",
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "size", "in": "query", "schema": {"type": "integer", "default": 50}},
          {"name": "type", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {"200": {"description": "List of nodes"}},
        "tags": ["read"]
      }
    },
    "/nodes/{node_id}": {
      "get": {
        "summary": "Get node by ID",
        "parameters": [{"name": "node_id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Node details"}, "404": {"description": "Node not found"}},
        "tags": ["read"]
      }
    },
    "/nodes/{node_id}/subgraph": {
      "get": {
        "summary": "Get node subgraph (IDs only)",
        "parameters": [
          {"name": "node_id", "in": "path", "required": true, "schema": {"type": "integer"}},
          {"name": "hops", "in": "query", "schema": {"type": "integer", "default": 1}}
        ],
        "responses": {"200": {"description": "Subgraph node and edge IDs"}},
        "tags": ["read"]
      }
    },
    "/nodes/{node_id}/subgraph/hydrated": {
      "get": {
        "summary": "Get hydrated subgraph",
        "description": "Get subgraph with full node and edge data",
        "parameters": [
          {"name": "node_id", "in": "path", "required": true, "schema": {"type": "integer"}},
          {"name": "hops", "in": "query", "schema": {"type": "integer", "default": 1}}
        ],
        "responses": {"200": {"description": "Hydrated subgraph with full data"}},
        "tags": ["read"]
      }
    },
    "/hyperedges": {
      "get": {
        "summary": "List hyperedges (paginated)",
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "size", "in": "query", "schema": {"type": "integer", "default": 50}}
        ],
        "responses": {"200": {"description": "List of hyperedges"}, "503": {"description": "Graph store not available"}},
        "tags": ["read"]
      }
    },
    "/hyperedges/{edge_id}": {
      "get": {
        "summary": "Get hyperedge by ID",
        "parameters": [{"name": "edge_id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Hyperedge details"}, "404": {"description": "Hyperedge not found"}},
        "tags": ["read"]
      }
    },
    "/contradictions": {
      "get": {
        "summary": "List contradiction hyperedges",
        "responses": {"200": {"description": "List of contradictions"}, "503": {"description": "Graph store not available"}},
        "tags": ["read"]
      }
    },
...
  },
...
}